fix: indexof issue
